+++ data/split_bar_separated.sh
@@ -1,1 +1,3 @@
-CQChartsTest -ceil -exec split_bar_separated.cl -close_app
+CQChartsTest -tcl -exec split_bar_separated.tcl -loop
+
+#CQChartsTest -ceil -exec split_bar_separated.cl -close_app