+++ data/var_plot.sh
@@ -1,0 +1,3 @@
+CQChartsTest -tcl -exec var_plot.tcl -loop
+
+#CQChartsTest -ceil -exec var_plot.cl