+++ include/CQChartsPlot.h
@@ -182,15 +182,6 @@ class CQChartsPlotType {
 
 //----
 
-class CQChartsHierPlotType : public CQChartsPlotType {
- public:
-  CQChartsHierPlotType();
-
-  void addParameters() override;
-};
-
-//----
-
 class CQChartsPlotUpdateTimer : public QTimer {
  public:
   CQChartsPlotUpdateTimer(CQChartsPlot *plot) :
@@ -275,6 +266,7 @@ class CQChartsPlot : public QObject {
   Q_PROPERTY(bool           logY                READ isLogY              WRITE setLogY           )
   Q_PROPERTY(bool           autoFit             READ isAutoFit           WRITE setAutoFit        )
   Q_PROPERTY(bool           preview             READ isPreview           WRITE setPreview        )
+  Q_PROPERTY(int            previewMaxRows      READ previewMaxRows      WRITE setPreviewMaxRows )
   Q_PROPERTY(bool           showBoxes           READ showBoxes           WRITE setShowBoxes      )
 
  public:
@@ -551,6 +543,11 @@ class CQChartsPlot : public QObject {
 
   bool isPreview() const { return preview_; }
   void setPreview(bool b) { preview_ = b; }
+
+  int previewMaxRows() const { return previewMaxRows_; }
+  void setPreviewMaxRows(int i) { previewMaxRows_ = i; }
+
+  //---
 
   bool showBoxes() const { return showBoxes_; }
   void setShowBoxes(bool b) { showBoxes_ = b; update(); }
@@ -1389,6 +1386,7 @@ class CQChartsPlot : public QObject {
   bool                      followMouse_      { true };       // track object under mouse
   bool                      autoFit_          { false };      // auto fit on data change
   bool                      preview_          { false };      // is preview plot
+  int                       previewMaxRows_   { 1000 };       // preview max rows
   bool                      showBoxes_        { false };      // show debug boxes
   bool                      invertX_          { false };      // x values inverted
   bool                      invertY_          { false };      // y values inverted
@@ -1412,69 +1410,6 @@ class CQChartsPlot : public QObject {
   CQChartsPlotUpdateTimer*  updateTimer_      { nullptr };    // update timer
   CQChartsEditHandles       editHandles_;                     // edit controls
   Annotations               annotations_;                     // extra annotations
-};
-
-//------
-
-class CQChartsHierPlot : public CQChartsPlot {
-  Q_OBJECT
-
-  Q_PROPERTY(CQChartsColumn nameColumn  READ nameColumn     WRITE setNameColumn    )
-  Q_PROPERTY(QString        nameColumns READ nameColumnsStr WRITE setNameColumnsStr)
-  Q_PROPERTY(CQChartsColumn valueColumn READ valueColumn    WRITE setValueColumn   )
-  Q_PROPERTY(CQChartsColumn colorColumn READ colorColumn    WRITE setColorColumn   )
-  Q_PROPERTY(QString        separator   READ separator      WRITE setSeparator     )
-
- public:
-  CQChartsHierPlot(CQChartsView *view, CQChartsPlotType *type, const ModelP &model);
-
- ~CQChartsHierPlot();
-
-  //---
-
-  const CQChartsColumn &nameColumn() const { return nameColumn_; }
-  void setNameColumn(const CQChartsColumn &c);
-
-  const Columns &nameColumns() const { return nameColumns_; }
-  void setNameColumns(const Columns &nameColumns);
-
-  QString nameColumnsStr() const;
-  bool setNameColumnsStr(const QString &s);
-
-  int numNameColumns() const { return nameColumns_.size(); }
-
-  const CQChartsColumn &valueColumn() const { return valueColumn_; }
-  void setValueColumn(const CQChartsColumn &c);
-
-  //---
-
-  const CQChartsColumn &colorColumn() const { return valueSetColumn("color"); }
-  void setColorColumn(const CQChartsColumn &c) {
-    setValueSetColumn("color", c); updateRangeAndObjs(); }
-
-  bool isColorMapped() const { return isValueSetMapped("color"); }
-  void setColorMapped(bool b) { setValueSetMapped("color", b); updateObjs(); }
-
-  double colorMapMin() const { return valueSetMapMin("color"); }
-  void setColorMapMin(double r) { setValueSetMapMin("color", r); updateObjs(); }
-
-  double colorMapMax() const { return valueSetMapMax("color"); }
-  void setColorMapMax(double r) { setValueSetMapMax("color", r); updateObjs(); }
-
-  //---
-
-  const QString &separator() const { return separator_; }
-  void setSeparator(const QString &s) { separator_ = s; }
-
-  //---
-
-  void addProperties() override;
-
- protected:
-  CQChartsColumn nameColumn_  { 0 };   // name column
-  Columns        nameColumns_;         // multiple name columns
-  CQChartsColumn valueColumn_;         // value column
-  QString        separator_   { "/" }; // hierarchical name separator
 };
 
 #endif